+++ packages/orq-rc/src/models/components/conversationwithmessagesresponse.ts
@@ -122,7 +122,7 @@ export const ConversationWithMessagesResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg1dfq22qq2ssjfvjb54dh09"),
+  _id: z.string().default("conv_01kg2rzqd2zp2khc3k3hexrtgd"),
   kind: ConversationWithMessagesResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),