+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-28T04:23:34.300Z",
+    "2026-01-28T17:03:47.548Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {