+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -180,7 +180,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg1dfq20hherk5t3k82e3mcx"),
+  _id: z.string().default("conv_01kg2rzqd06dhk2gje72axp614"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),